refactor: move Miner and MinerWatcher to skepticoin/mining.py
